Add JS files.
